some adjustment for cygwin faster script when downloading

+++ data/peaks/scripts/download_and_process_one.sh
@@ -1,30 +1,49 @@
 cd ../raw_bed
 
 wget -O $1-H3K27ac.narrowPeak.gz "https://egg2.wustl.edu/roadmap/data/byFileType/peaks/consolidated/narrowPeak/$1-H3K27ac.narrowPeak.gz"
+if [ ! -f $1-H3K27ac.narrowPeak.gz ]; then
+    echo "$1-H3K27ac.narrowPeak.gz not found!"
+    exit 1
+fi
 wget -O $1-H3K4me3.narrowPeak.gz "https://egg2.wustl.edu/roadmap/data/byFileType/peaks/consolidated/narrowPeak/$1-H3K4me3.narrowPeak.gz"
+if [ ! -f $1-H3K4me3.narrowPeak.gz ]; then
+    echo "$1-H3K4me3.narrowPeak.gz not found!"
+    exit 1
+fi
 wget -O $1-H3K4me1.narrowPeak.gz "https://egg2.wustl.edu/roadmap/data/byFileType/peaks/consolidated/narrowPeak/$1-H3K4me1.narrowPeak.gz"
+if [ ! -f $1-H3K4me1.narrowPeak.gz ]; then
+    echo "$1-H3K4me1.narrowPeak.gz not found!"
+    exit 1
+fi
 wget -O $1-H3K27me3.narrowPeak.gz "https://egg2.wustl.edu/roadmap/data/byFileType/peaks/consolidated/narrowPeak/$1-H3K27me3.narrowPeak.gz"
+if [ ! -f $1-H3K27me3.narrowPeak.gz ]; then
+    echo "$1-H3K27me3.narrowPeak.gz not found!"
+    exit 1
+fi
 
+echo "done downloading"
 gunzip -f ./*.gz
 rm -f ./*.gz
+echo "done extracting"
 
 
 if [ ! -f $1-H3K27me3.narrowPeak ]; then
-    echo "File not found!"
+    echo "$1-H3K27me3.narrowPeak not found!"
     exit 1
 fi
 if [ ! -f $1-H3K27ac.narrowPeak ]; then
-    echo "File not found!"
+    echo "$1-H3K27ac.narrowPeak not found!"
     exit 1
 fi
 if [ ! -f $1-H3K4me1.narrowPeak ]; then
-    echo "File not found!"
+    echo "$1-H3K4me1.narrowPeak not found!"
     exit 1
 fi
 if [ ! -f $1-H3K4me3.narrowPeak ]; then
-    echo "File not found!"
+    echo "$1-H3K4me3.narrowPeak not found!"
     exit 1
 fi
+echo "asserted files existing"
 
 # refine with background
 bedtools sort -faidx hg19.chrom.sizes -i $1-H3K27me3.narrowPeak > tmp.1.narrowPeak
@@ -44,11 +63,13 @@ bedtools complement  -i tmp.1.narrowPeak -g hg19.chrom.sizes > tmp.2.narrowPeak
 bedtools intersect -a tmp.2.narrowPeak -b background.narrowPeak > tmp.3.narrowPeak
 mv -f tmp.3.narrowPeak background.narrowPeak
 rm -f tmp.*
+echo "bedtools run on background completed"
 
 if [ ! -f background.narrowPeak ]; then
-    echo "File not found!"
+    echo "background.narrowPeak not found!"
     exit 1
 fi
+echo "asserted background update"
 
 #take only H3k27ac peaks with H3k4me1 peaks
 bedtools intersect -a $1-H3K27ac.narrowPeak -b $1-H3K4me1.narrowPeak -wa >$1-H3K27ac.cleaned.narrowPeak
@@ -62,6 +83,7 @@ bedtools intersect -a $1-H3K27ac.narrowPeak -b $1-H3K4me1.narrowPeak -wa >$1-H3K
 awk '!x[$0]++' $1-H3K27ac.cleaned.narrowPeak > $1-H3K27ac.tmp.narrowPeak
 mv -f $1-H3K27ac.tmp.narrowPeak $1-H3K27ac.cleaned.narrowPeak
 
+echo "duplicates removed"
 
 #remove H3k27me3 peaks
 bedtools subtract -a $1-H3K27ac.cleaned.narrowPeak -b $1-H3K27me3.narrowPeak -A > $1-H3K27ac.tmp.narrowPeak
@@ -76,8 +98,9 @@ bedtools subtract -a $1-H3K27ac.cleaned.narrowPeak -b ./hg19.KnownGenes.slopped.
 mv -f $1-H3K27ac.tmp.narrowPeak $1-H3K27ac.cleaned.narrowPeak
 mv -f $1-H3K27ac.cleaned.narrowPeak ../processed/$1-H3K27ac.cleaned.narrowPeak
 
+echo "bedtools run completed"
 
 if [ ! -f ../processed/$1-H3K27ac.cleaned.narrowPeak ]; then
-    echo "File not found!"
+    echo "../processed/$1-H3K27ac.cleaned.narrowPeak not found!"
     exit 1
 fi
